Add files via upload
Built and Balance Search Tree data structure in JS where we are able to look up / seek it's value
Also built a queue with BFS algo. 
with the BFS algo we now can get an array with a list of data that matches the Balanced Tree data structure as well.